# src/backend/docs/@typespec/openapi3/openapi.v1.yaml
openapi: 3.0.0
info:
  title: RacketShop
  version: v1
tags:
  - name: auth
paths:
  /api/auth/login:
    post:
      operationId: Auth_login
      summary: login user
      description: Login user
      parameters: []
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        default:
          description: An unexpected error response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      tags:
        - auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
              required:
                - email
                - password
  /api/auth/register:
    post:
      operationId: Auth_register
      summary: register user
      description: Register user
      parameters: []
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterResponse'
        default:
          description: An unexpected error response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      tags:
        - auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                name:
                  type: string
                password:
                  type: string
                surname:
                  type: string
              required:
                - email
                - name
                - password
                - surname
components:
  schemas:
    ErrorResponse:
      type: object
      required:
        - type
      properties:
        type:
          $ref: '#/components/schemas/Types'
    LoginRequest:
      type: object
      required:
        - loginBody
      properties:
        loginBody:
          type: object
          properties:
            email:
              type: string
            password:
              type: string
          required:
            - email
            - password
    LoginResponse:
      type: object
      allOf:
        - $ref: '#/components/schemas/ResponseWithAccessToken'
    RegisterRequest:
      type: object
      required:
        - registerBody
      properties:
        registerBody:
          type: object
          properties:
            email:
              type: string
            name:
              type: string
            password:
              type: string
            surname:
              type: string
          required:
            - email
            - name
            - password
            - surname
      description: API Requests
    RegisterResponse:
      type: object
      allOf:
        - $ref: '#/components/schemas/ResponseWithAccessToken'
      description: API Responses
    ResponseWithAccessToken:
      type: object
      required:
        - access_token
      properties:
        access_token:
          type: string
    Types:
      type: string
      enum:
        - BadRequest
        - ValidationError
        - NotFoundError
        - InternalError
        - NotModifiedError
        - ConflictError
    Versions:
      type: string
      enum:
        - v1
servers:
  - url: http://localhost:8080
    description: localhost
    variables: {}